Filter ext4 features and options more rigorously
Subshell <() redirection is unavailable early in the initrd
so we need to use $() constructs instead when reading fs info.

Not all "options" reported by tune2fs et al are actually settable
even though they are listed as features: Use an allowlist to make
sure we don't try to set a non-feature (which will fail).