steamos-update-os: Make the script a wrapper around steamos-update
`steamos-update-os` is a lefover script from the early development days.
Compared to the canonical `steamos-update`, this is simply calling
`steamos-atomupd-client` with no additional logic.

Apparently this script was introduced to allow automatic system updates
in background, e.g. at night when the system was not in use. This idea
was then discarded because it was not suited for handheld devices.

At the moment this script doesn't seem to be in use anymore, nor with
the current Jupiter stable (20230508.1) nor main (20230612.1000).
But to avoid a potential disruption, with this commit we add a
deprecation notice and make it call the canonical `steamos-update`
instead.
In this way we ensure a consistent experience without the burden to
support two scripts that have the same goal.

Once we are absolutely certain that we don't need "steamos-update-os"
anymore, we can proceed to completely remove it.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>