mkinitcpio: drop alpm hook, use config snippet
The hook was meant to change the config file. This is no longer needed
with newer mkinitcpio, which supports config snippets.

This will allow us to have unchanged mkinitcpio.conf file and less of a
reason for people to change it. Hence less of a headache when dealing
with our /etc overlay.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>