Revert "atomic-update: exclude from top-level makefile"
This reverts commit b021ef4fdeecca9924279db32864a05311e3a134.

The commit requires changes to the PKGBUILDs. While those are ready,
they do require fixes to our CI, that need proper testing first.

There was a tag & uprev that landed recently breaking the update for
everyone, since it lacked the PKGBUILD bits.

Revert this for now until everything is ready and we can merge both
sides in one go.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>