steamos-readonly: do not require root to get "status"
The rootfs was changed to BTRFS, so we no longer need to run tune2fs as
root to get status. With that change, getting "status" in this script no
longer requires root.

This change was done in a way that allows some actions to opt-out of
making root a requirement for execution.

As a tiny bonus, this also fixes a quoting error from shellcheck (
when doing "basename $0").