desync: Tweak the error retry interval
500ms of base wait is the default value that is expected to be provided
by the upcoming Desync 0.9.6.

Lowering the wait time from 1s to 500ms can speed up the update process
when the connection is only slightly unstable and drops a few packages
along the way.

In order to allow the connection to be down for a reasonable consecutive
amount of time, we also rise the number of retries before Desync gives
up.

With 35 retries and a base interval of 500ms, the total amount of wait
time will be (0.5*35+0.5)*35/2 = 315 seconds (~5 minutes)

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>