atomic-update: Abort an update if we can't sync /var
By using `sync_var_mountpoints [...] || err "Failed to sync [...]"` the
intention was to error out if rsync failed to sync the current `/var`
partition with the target.

However, `set -e` doesn't abort the execution if `sync_var_mountpoints`
is executed with the OR operator. This means that even if `rsync`
failed, we unexpectedly never reached the
`err "Failed to sync var partitions"`.

Instead, we wrap `rsync` inside an `if` to return an error code if
`rsync` fails.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>