atomic-update: Create a backup with the /etc changes
Given that we keep only the strictly necessary changes from /etc, it
might be useful to backup the changes too.

For example, they may allow users to retrieve their `/etc` changes if
they forgot to list them in the drop-in config directory
`/etc/atomic-update.conf.d/`

Under normal circumstances, the backups are expected to be quite small
in size (just a few KB).

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>