Revert "rauc: Enable Desync"
This reverts commit c553fbbdd22ed2820a24db00027fdf646fd53a3a.

If we want to use development branches/CI branches, Desync requires
the usage of the new `client-dev.conf` file, instead of simply executing
`steamos-select-branch --ci-env`.
However, `client-dev.conf` is not yet in use, so switching to Desync at
this very moment will likely break the update system for the devices
that are currently tracking CI branches.
We postpone this change until `client-dev.conf` actually starts to be
used.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>