atomic-update: Keep only the strictly necessary changes from /etc
Re-apply the commit fa35533cc31d9aecc447a2b031b52c83e289e1cf that was
reverted by e5688710371126471c31976232313fd4de361544.

The problem we encountered was that in some situations the `/etc` changes
were not synchronized to the destination.
Apparently this issue was not appearing when the readonly mode was
disabled.

It turned out that the issue was in how we created the temporary config
file. In order to create a file under `/tmp` we need to explicitly use
the flag `--tmpdir` with `mktemp`. Otherwise `mktemp` was attempting to
create the file in the same working directory, which could have been RO.

The diff between the old fa35533cc31d9aecc447a2b031b52c83e289e1cf
is the following:
```
-    config=$(mktemp "etc_config.XXXXXX")
+    config=$(mktemp --tmpdir "etc_config.XXXXXX")
+    if [ -z "${config}" ]; then
+        warn "Failed to create a temporary file for rsync"
+        return 1
+    fi
```

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>