atomic-update: Remove old query URL for the dynamic server
In Jupiter we switched to the new static meta JSON server in all
branches (from main up to stable).

`steamos-atomupd` uses the new `MetaUrl` as its first choice, and falls
back to the old `QueryUrl` in case the server is unreachable.
This design allowed us to switch to the new server, while keeping the
old one around to ease the transition.

With jupiter/tasks#448 we are planning on stopping the deprecated
dynamic server, and redirecting the query requests to the generic static
response `steamdeck.json` instead.
Before this happens, we should remove `QueryUrl` for new Jupiter images.
Because, by then, the canonical `MetaUrl` response might be different
from the old `QueryUrl`, and we don't want to propose different updates
if users are experiencing a flaky Internet connection.
This is especially true when we'll release out first checkpoint, because
in that situation the generic `steamdeck.json` will point to an
unsupported update.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>