Check EFI partition for errors
We expect the EFI partition to be VFAT, because this is the supported
filesystem under UEFI. If it is VFAT, it is unfortunately not
resilient to errors, and in particular we can get repeated (and hard
to diagnose) boot failures if we apply updates onto a VFAT partition
that has critical errors.

This commit is the start of trying to be more resilient to broken VFAT
filesystems. We abort if the EFI filesystem is of the wrong type,
unless it's exFAT (because exFAT will work, and the user had to
manually opt for that option). We don't accept any other filesystem
type, including NTFS. In later commits we'll fix partitions with
inappropriate filesystems by reformatting them.

Signed-off-by: Ed Smith <ed.smith@collabora.com>