Check ESP partition for errors
There is only a single ESP partition (rather than one per slot), but
as for the EFI partition, we'd like to check it for errors on
install. As for the EFI partitions, we expect it to be VFAT, can
tolerate it being exFAT, and anything else is unexpected and
unacceptable for the stability.

Signed-off-by: Ed Smith <ed.smith@collabora.com>