Rebuild the EFI partition
If fsck detects unfixable errors, we can do better than simply
abort. Instead, we can format the affected EFI partition and then
restore them by running `steamos-partsets`. We own this partition, and
its contents are completely reproducible from the read-only
filesystem, so this is a good way to avoid having a unit that cannot
be upgraded.

Signed-off-by: Ed Smith <ed.smith@collabora.com>