steamos-partsets: Remove unnecessary code
First we append the link data to "$OUTDIR/$partset". We're inside a
`case` statement, where the specific condition we're in is
`$partset == $SELF`. So this could equally have been an echo to
`$OUTDIR/$SELF`.

The added stanza checks for "$SELF == dev", and then appends link data
to "$OUTDIR/dev" which is the same as "$OUTDIR/$SELF" (because of the
test), which is the same as the previous line. Thus the added stanza
for the dev case is unnecessary.

Signed-off-by: Ed Smith <ed.smith@collabora.com>