Remove mkswapfile, add steamos-rm-default-swapfile
Change home-swapfile.swap unit to enable swapfile as active swap area
only if the file exists.

Swapping (pun intended?) command 'mkswapfile' and corresponding
one-shot type of service to execute on boot to their effective reverse
operation, 'rm-swapfile'.  So instead of creating a swapfile on every
boot, attempt to remove deprecated /home/swapfile.

The swapfile is only removed under a variety of conditions, for safey,
like:
* unused at the time of execution
* it's actually a swapfile
* it matches the properties like default size with which older SteamOS
  releases created it

In summary, if it's the default swapfile created by SteamOS in past
releases.

If the conditions are not met, the assumption is that some kind of
human intervention happened in the meantime, which means that most
probably the swapfile should stay that way and not removed.
E.g. users enabling larger swap area because some apps/games need it.

So, in those cases, playing it safe by leaving the swapfile in place.