misc: sysctl.d: Prevent split lock slowdown on SteamOS
Split locks on user applications started to incur a strong artificially
imposed slowdown on the applications that caused them, as a punishment
from the kernel - this behavior was introduced on kernel version 5.19.
There's a sysctl introduced in kernels 6.0+ that disables this behavior
though (details on [0]).

We should tune such sysctl on SteamOS since games could experience split
locks, and given that they're (usually) proprietary software, it's not easy
nor quick to have them fixed. Better to prevent the big slowdown imposed
by the kernel then.

Notice it shouldn't be a problem for Steam Deck since AMD CPUs don't
have a split lock detector, but it could in the future (and also
*it is* a problem now for many Intel CPUs).

Gitlab issue: https://www.youtube.com/watch?v=WxnN05vOuSM [1]

[0] https://lwn.net/Articles/911219/

[1] https://gitlab.steamos.cloud/jupiter/tasks/-/issues/666

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>