Adding 'steamos-set-wifi-backend'
The program is a helper to change NetworkManager configuration, and
can:

- with "--check": detect and print the current "wifi.backend"

- otherwise, if given the name of a back-end as argument ("iwd" or
  "wpa_supplicant"), switch between these back-ends

The intention with this helper program is to make it easy to switch
back-ends, which is useful in the case that one of the back-ends does
not support some features, or is not allowing connections for some
reason.

Doing it in this way is manageable by inexperienced users where
otherwise it might not be, and more convenient and faster for more
experienced users.