atomic-update: Preserve /etc/environment by default
System-wide environment variables are usually set by editing the
`/etc/environment` file.

It is very unlikely that we'll ever release an image that relies on a
particular environment variable set in `/etc/environment`.
For this reason, the risk of user edits shadowing an image provided file
are rather slim.

We include that file in the default keep list to make it easier for end
users to set a custom environment variable on their system.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>