steamos-chroot: Replace additional forward slashes from the lock name
When using the `--disk` option, it might happen that the `dev_name`
contains one or more forward slashes.
We should replace them to avoid attempting to create a lock in a
subdirectoy that doesn't exist.

This fixes the following error when using `--disk /dev/nvme0n1`:
```
/usr/bin/steamos-chroot: line 206: /run/lock//dev/nvme0n1-A.flock: No
such file or directory
```

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>