desync: Do not hardcode the `steamdeck` directory name
When creating a development atomupd configuration file,
`atomupd-manager` appends the generic entry `${URL}/*/...`, to ensure
the maximum compatibility with the server URL layout.
But, when attempting to keep the existing desync config in sync, the
hard coded `${URL}/steamdeck/...` will inevitably be treated as a
different path.

This could result in a config error because Desync doesn't
allow multiple entries to match a single location.

For this reason we remove the explicit `steamdeck` to match what
`atomupd-manger` uses.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>