modules-load.d: Preload hid_playstation.ko
When a DualShock 4 or DualSense is plugged in the first time during a boot,
there exists a race condition between Steam and the hid_playstation driver
trying to claim the device. This can lead to the hidraw disappearing and Steam
falling back to the evdev node, which has degraded functionality. Due to a bug
somewhere in the bluetooth and/or input stack, it can also fail to disappear
from the hid subsystem when the device is removed, too. By preloading the
module this race condition can be avoided.