atomic-update: Prevent the /etc overlay from masking new files in the dest rootfs
An /etc/foo directory deleted by the user is represented in the upper
dir of the /etc overlay with a special "whiteout" file, which means:
this entry is deleted, ignore the contents of the lowerdir.

If this whiteout file is copied to the new overlay during an OS update
the underlying /etc/foo directory will remain deleted. This means that
if the new rootfs adds new files to that directory they won't be
visible by the user.

This can be a problem for systemd activation symlinks: if the new
rootfs comes with a new service that is enabled by default it won't be
visible if the '*.wants' directory where the activation symlink goes
was deleted by the user before the OS update.

This can be solved by replacing deleted directories in the upper dir
with actual directories where each one of the original contents is
marked as deleted.

A related problem happens if a directory is marked as opaque in the
upper dir: since the opaque attribute is lost during an OS update we
need to mark the original contents as deleted to make sure that they
don't reappear (keeping the opaque attribute is not a solution either
because we'd run into the whiteout problem described earlier).

This patch only applies this solution to /etc/systemd, where this
problem can happen easily and the solution is straightforward. Other
cases would need to be analyzed individually.

Fixes https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1204

Signed-off-by: Alberto Garcia <berto@igalia.com>