atomic-update: Preserve the provided exit value in post-install script
When using the `fail()` function, we run `exit 1` without explicitly set
the $ERR variable first.

This means that if in `cleanup()` we don't save `$?`, the last `exit
$ERR` might unexpectedly return zero as the error code.

To also cover any potential `exit ${CUSTOM_EXIT_CODE}` that we may add
in the future, we save the error code in `cleanup()` instead of using
`ERR=1` in `fail()`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>