atomic-update: Handle the image activation in the post-install
By default, when RAUC starts to apply an update, it runs `set-state
${other} bad` to mark that partition as unusable.
When the update is completed, it then runs `set-state ${other} good`.

The problem is that RAUC considers the update to be completed after the
install step, and not after the post-install script.
So, regardless of what happens in post-install.sh, after a reboot the
user will go to the other partition.

By using the `activate-installed=false` option we can instruct RAUC to
skip the `set-state ${other} good` part. Then, we will mark that
partition as valid ourselves at the end of post-install.sh.

In post-install we have to change the `--set image-invalid 0` to
`set-mode reboot`, because we also need to clear the `boot-other 1`
flag. `set-mode reboot` already takes care of setting the
`image-invalid` flag to 0, so there is no need to do that explicitly.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>
(cherry picked from commit 941e2e84917e717a83f50eac47bc78b5bcee76f1)