NetworkManager: Clean up config and move to /usr
This moves all of the SteamOS-specific NetworkManager configuration into a
single file. Since NetworkManager loads files in lexicographical order, we use
a numbered prefix for the file name so it can be overridden with files that
don't need creative prefixing. Since these files aren't meant to be
user-editable this moves it to the /usr version of the config directory. Values
can still be overridden by placing files in /etc if needed.