atomic-update: Clean the pacman cache after an OS update
This new 'steamos-post-update.service' runs on the first boot after
an OS update and it currently cleans the pacman package cache. This
solves a problem where trying to install packages after an update
fails due to invalid signatures.

The /var/lib/steamos-atomupd/system-updated file which triggers this
new service is created by steamos-finalize-install.

Closes: https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1440