atomic-update: Install client.conf in /usr/lib and symlink it from /etc
`client.conf` is a fundamental file for atomupd. In the unusual use case
where end users want to change it, they should create a
`client-dev.conf` file instead of editing the `client.conf` that we ship
with SteamOS.

We had a few cases in the field where users unexpectedly edited the
`client.conf` file and ended up breaking the atomic updates.

In order to make the system more resilient, we now install the
`client.conf` in the RO `/usr/lib` partition and make
`/etc/steamos-atomupd/client.conf` a symlink.

`atomupd-daemon` will then be able to fallback to the
`/usr/lib/steamos-atomupd/client.conf` file, in case the canonical
`/etc` one is misconfigured.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>