Streamline the /etc overlay by passing the x-initrd.mount option
The x-initrd.mount option ensures that the overlay is mounted early in the
initrd. There was an old concern that this wouldn't work because upperdir and
workdir weren't guaranteed to exist at first boot, but since systemd 256
systemd.mount ensures that they are created if they don't exist.

Fixes holo-team/tasks#1510.