Use new functionality in std 'mkswap' instead of shipping ad-hoc 'mkswapfile'
Since this ad-hoc program was created and shipped by SteamOS, 'mkswap'
gained the ability to create swapfiles in the right way, with the
right permissions and so on, and "formatting" them as necessary for
use as swap.

So drop our ad-hoc utility program and use the new standard way
instead.