steamos-finalize-install: Don't fail if we can't create the trigger file
If for whatever reason we can't create the trigger file
`/var/lib/steamos-atomupd/system-updated`, it's better to log a warning
and continue with the update.

Even if this will result in not running `steamos-post-update.service` on
first boot, the operations are not critical and the system will continue
to work as expected.

Part of https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1803
Helps with https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1833

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>