steamos-finalize-install: Ensure the directory exist before creating a file
We operated under the assumption that `/var/lib/steamos-atomupd` was
guaranteed to be present, because that was a mandatory operation from
`post-install.sh`.

However, if you are applying an update from a very old OS image, the
existence of that directory may not be a guarantee anymore.

With this commit we manually ensure that the directory exists before
proceeding with the `touch` command.

Addresses https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1833

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>
(cherry picked from commit 33fe7c2e4f4f56c4b5710013d9bf1fcf7115c32b)