atomic-update: Add trusted_keys directory with symlink to keyring.pem
As part of holo-team/tasks#1129, we introduce a directory that will hold
all keys that need to be trusted by RAUC.

With this commit we create that directory with a symlink for the
canonical `keyring.pem` file, that gets provided by holo/holo during the
image creation step.

The name of the symlink must follow the convention of
X509_LOOKUP_hash_dir(3).
To generate it, it was used:
`openssl x509 -subject_hash -noout -in keyring.pem`

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>