holo-sync-var: Exclude /etc/previous and etc_backup in catchup mode
When running this script in `catchup` mode, we are not going to create
new backups of `/etc`.

For this reason, we need to preserve `/etc/previous` and
`/var/lib/steamos-atomupd/etc_backup` from the destination.
Otherwise `/etc/previous` will end up being wiped after
the rsync, and `etc_backup` will miss the latest backup.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>